+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 
 
+## [0.12.0](https://github.com/Blobfolio/dactyl/releases/tag/v0.12.0) - TBD
+
+### Changed
+
+* `Nice*::as_bytes` is now const
+* `Nice*::as_str` is now const
+* `NiceFloat::compact_bytes` is now const
+* `NiceFloat::compact_str` is now const
+* `NiceFloat::precise_bytes` is now const
+* `NiceFloat::precise_str` is now const
+
+
 ## [0.11.1](https://github.com/Blobfolio/dactyl/releases/tag/v0.11.0) - 2025-05-30
 
 ### Changed